strlcpy changes, still doesnt pass with null

+++ ft_strlcpy.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mpeterso <mpeterso@student.42madrid.com>   +#+  +:+       +#+        */
+/*   By: mpeterso <mpeterso@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/22 12:01:07 by mpeterso          #+#    #+#             */
-/*   Updated: 2022/10/07 13:45:32 by mpeterso         ###   ########.fr       */
+/*   Updated: 2022/10/25 19:59:46 by mpeterso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	unsigned int	i;
-	unsigned int	srclen;
+	size_t	i;
+	size_t	srclen;
 
 	if (dst == NULL || src == NULL)
 	{
